requirements.txt
absl-py==2.0.0
black==23.10.1
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.9.0
fonttools==4.43.1
fsspec==2023.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
isort==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.5
Markdown==3.5
MarkupSafe==2.1.2
matplotlib==3.7.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
packaging==23.2
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.11.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
zipp==3.17.0
